pyproject.toml
[project]
name = "email-list-manager"
version = "0.1.0"
description = "Python tools for managing and cleaning email lists from Sendy database"
authors = [
    {name = "Cody O'Rourke"}
]
readme = "README.md"
requires-python = "^3.8"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
